Fix missing cast to int in Fox GetNutritionalValue

+++ Code/Ecosystem Simulation/Assets/Scripts/Animals/Fox.cs
@@ -42,6 +42,6 @@ public class Fox : Animal
 
     public override int GetNutritionalValue()
     {
-        return baseNutritionalValue * age;
+        return baseNutritionalValue * (int)age;
     }
 }
